requirements.txt
python-telegram-bot>=21.0
imghdr
flask==2.3.3
python-dotenv==1.0.0
